+++ packages/react-appkit/src/components/PartySharingDialog.tsx
@@ -2,7 +2,7 @@
 // Copyright 2020 DXOS.org
 //
 
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { CopyToClipboard } from 'react-copy-to-clipboard';
 
 import { makeStyles, withStyles } from '@material-ui/core';
@@ -34,7 +34,7 @@ import { BotFactoryClient } from '@dxos/botkit-client';
 import { keyToBuffer, verify, SIGNATURE_LENGTH } from '@dxos/crypto';
 import { useClient, useContacts, useInvitation, useOfflineInvitation } from '@dxos/react-client';
 
-import { useMembers } from '../hooks';
+import { useMembers, useSentry } from '../hooks';
 import BotDialog from './BotDialog';
 import MemberAvatar, { getAvatarStyle } from './MemberAvatar';
 import { Contact, Party, PartyMember } from '@dxos/echo-db';
@@ -205,30 +205,104 @@ const PartySharingDialog = ({
   const [copiedSnackBarOpen, setCopiedSnackBarOpen] = useState(false);
 
   const members: PartyMember[] = useMembers(party);
+  const [botFactoryTopic, setBotFactoryTopic] = useState<string | null>();
+  const [botFactoryClient, setBotFactoryClient] = useState<BotFactoryClient | null>();
+  const [botFactoryConnected, setBotFactoryConnected] = useState(false);
+  const [botToSpawn, setBotToSpawn] = useState<string | null>();
+  const [botInvitationPending, setBotInvitationPending] = useState(false);
+  const [botInvitationError, setBotInvitationError] = useState();
+  const [inviteRequestTime, setInviteRequestTime] = useState<number>();
+  const sentry = useSentry();
+
   const [contacts] = useContacts();
   const invitableContacts = contacts?.filter(c => !members.some(m => m.publicKey.toHex() === c.publicKey.toHex())); // contacts not already in this party
 
   const createInvitation = () => setInvitations([{ id: Date.now() }, ...invitations]);
-  const createOfflineInvitation = (contact: Contact) => setContactsInvitations(old => [...old, { id: Date.now(), contact }]);
 
-  const handleBotInvite = async (botFactoryTopic: string, botId: string | undefined, spec: Record<string, unknown> = {}) => {
-    const botFactoryClient = new BotFactoryClient(client.echo.networkManager, botFactoryTopic);
+  // const handleBotInvite = async (botFactoryTopic: string, botId: string | undefined, spec: Record<string, unknown> = {}) => {
+    // const botFactoryClient = new BotFactoryClient(client.echo.networkManager, botFactoryTopic);
 
-    const secretProvider: any = () => null;
+  const createOfflineInvitation = (contact: Contact) => {
+    setContactsInvitations(old => [...old, { id: Date.now(), contact }]);
+    if (sentry) {
+      sentry.addBreadcrumb({ message: `Created offline invitation for contact: '${contact.displayName || 'Unknown'}', ${contact.publicKey.toHex()}` });
+      sentry.captureMessage('Offline invitation initiated.');
+    }
+  };
 
-    // Provided by inviter node.
-    const secretValidator = async (invitation: any, secret: Buffer) => {
-      const signature = secret.slice(0, SIGNATURE_LENGTH);
-      const message = secret.slice(SIGNATURE_LENGTH);
-      return verify(message, signature, keyToBuffer(botFactoryTopic));
-    };
+  const handleBotInviteClick = () => {
+    setBotDialogVisible(true);
+    setBotInvitationError(undefined);
+  };
 
-    const invitation = await party.createInvitation({ secretValidator, secretProvider });
+  const handleBotInvite = async (botId: string | undefined, spec: Record<string, unknown> = {}) => {
+    setBotInvitationError(undefined);
+    setBotInvitationPending(true);
+    try {
+      const secretProvider: any = () => null;
 
-    const botUID = await botFactoryClient.sendSpawnRequest(botId);
-    assert(botUID);
-    await botFactoryClient.sendInvitationRequest(botUID, party.key.toHex(), spec, invitation.toQueryParameters());
+      // Provided by inviter node.
+      const secretValidator = async (invitation: any, secret: Buffer) => {
+        const signature = secret.slice(0, SIGNATURE_LENGTH);
+        const message = secret.slice(SIGNATURE_LENGTH);
+        assert(botFactoryTopic);
+        return verify(message, signature, keyToBuffer(botFactoryTopic));
+      };
+
+      const invitation = await party.createInvitation({ secretValidator, secretProvider });
+
+      assert(botFactoryClient);
+      const botUID = await botFactoryClient.sendSpawnRequest(botId);
+      assert(botUID);
+      await botFactoryClient.sendInvitationRequest(botUID, party.key.toHex(), spec, invitation.toQueryParameters());
+      setBotDialogVisible(false);
+      // TODO(egorgripasov): Replace 2 calls above once protocol changes are on kube.
+      // await botFactoryClient.sendSpawnAndInviteRequest(botId, party.key.toHex(), invitation.toQueryParameters(), {});
+      setBotInvitationPending(false);
+      await handleBotInvitationEnd();
+    } catch (err) {
+      console.error(err);
+      setBotInvitationPending(false);
+      setBotInvitationError(err);
+    }
+  };
+
+  
+
+
+
+
+  const handleBotFactorySelect = async (topic: string, force?: boolean) => {
+    if (botFactoryTopic !== topic || force) {
+      setBotFactoryConnected(false);
+      setBotFactoryTopic(topic);
+      if (botFactoryTopic && botFactoryClient) {
+        await botFactoryClient.close();
+      }
+      const bfClient = new BotFactoryClient(client.echo.networkManager, topic);
+      setBotFactoryClient(bfClient);
+
+      const bfConnected = await bfClient.connect();
+      if (bfConnected) {
+        setBotFactoryConnected(true);
+      }
+    }
+  };
+
+  const handleBotInvitationEnd = async () => {
+    if (botFactoryClient) {
+      await botFactoryClient.close();
+    }
+    setBotFactoryClient(null);
+    setBotFactoryTopic(null);
+    setBotToSpawn(null);
+    setBotFactoryConnected(false);
     setBotDialogVisible(false);
+  };
+
+  const handleSubmit = (bot: string | undefined) => {
+    setInviteRequestTime(Date.now());
+    setBotToSpawn(bot);
   };
 
   const handleCopy = (value: string) => {
@@ -242,6 +316,12 @@ const PartySharingDialog = ({
       ...old.filter(invite => invite.id === invitationId).map(invite => ({ ...invite, done: true }))
     ]));
   };
+
+  useEffect(() => {
+    if (botFactoryConnected && botFactoryTopic && botToSpawn) {
+      handleBotInvite(botToSpawn);
+    }
+  }, [botFactoryConnected, botFactoryTopic, botToSpawn, inviteRequestTime]);
 
   // TODO(burdon): Columns in EACH section should have same content:
   // [SMALL AVATAR] [NAME] [INVITATION PIN] [MEMBER TYPE] [ACTIONS: e.g., refresh PIN/remove]
@@ -272,7 +352,7 @@ const PartySharingDialog = ({
             </Button>
             <Button
               size='small'
-              onClick={() => setBotDialogVisible(true)}
+              onClick={() => handleBotInviteClick()}
             >
               Invite Bot
             </Button>
@@ -281,8 +361,11 @@ const PartySharingDialog = ({
 
         <BotDialog
           open={botDialogVisible}
-          onSubmit={async ({ topic: bfTopic, bot, spec }) => handleBotInvite(bfTopic, bot, spec)}
-          onClose={() => setBotDialogVisible(false)}
+          invitationPending={botInvitationPending}
+          invitationError={botInvitationError}
+          onBotFactorySelect={async (bfTopic, force) => handleBotFactorySelect(bfTopic, force)}
+          onSubmit={async ({ bot }) => handleSubmit(bot)}
+          onClose={() => handleBotInvitationEnd()}
         />
 
         <Snackbar