+++ packages/react-appkit/src/containers/AppKitProvider.tsx
@@ -12,6 +12,8 @@ import errorsReducer, { SET_ERRORS } from '../hooks/errors';
 import filterReducer, { SET_FILTER } from '../hooks/filter';
 import layoutReducer, { SET_LAYOUT } from '../hooks/layout';
 
+const noop = () => {}; // eslint-disable-line
+
 const defaultState = {
   [SET_LAYOUT]: {
     showSidebar: true,
@@ -43,13 +45,15 @@ export interface AppKitProviderProps {
   errorHandler?: any,
   issuesLink?: string,
   router?: any,
+  keywords?: string[], // Used for sorting relevant registry records first
+  sentry?: any,
 }
 
 /**
  * Creates the AppKit framework context, which provides the global UX state.
  * Wraps children with a React ErrorBoundary component, which catches runtime errors and enables reset.
  */
-const AppKitProvider = ({ children, initialState, router = DefaultRouter, errorHandler, pads = [], issuesLink = undefined }: AppKitProviderProps) => {
+const AppKitProvider = ({ children, initialState, router = DefaultRouter, errorHandler, pads = [], issuesLink = undefined, keywords = [], sentry = undefined }: AppKitProviderProps) => {
   const client = useClient();
   const [state, dispatch] = useReducer(appReducer, defaultsDeep({}, initialState, defaultState));
   const [padsRegistered, setPadsRegistered] = useState(false);
@@ -82,8 +86,13 @@ const AppKitProvider = ({ children, initialState, router = DefaultRouter, errorH
     registerPadModels();
   }, []);
 
+  const profile = client.getProfile();
+  if (profile !== undefined && sentry !== undefined) {
+    sentry.setUser({ username: `${profile.username}-${profile.publicKey.toHex()}`, id: profile.publicKey.toHex() });
+  }
+
   return (
-    <AppKitContext.Provider value={{ state, dispatch, router, pads, issuesLink }}>
+    <AppKitContext.Provider value={{ state, dispatch, router, pads, issuesLink, keywords, sentry }}>
       {padsRegistered && children}
     </AppKitContext.Provider>
   );